requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
Flask-JWT-Extended==4.5.2
marshmallow==3.19.0